+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.5](https://github.com/tenseijs/tensei/compare/v0.11.4...v0.11.5) (2022-01-06)
+
+**Note:** Version bump only for package @tensei/main
+
+
+
+
+
+## [0.11.4](https://github.com/tenseijs/tensei/compare/v0.11.3...v0.11.4) (2022-01-06)
+
+
+### Features
+
+* Email display in team members table ([#222](https://github.com/tenseijs/tensei/issues/222)) ([b5595be](https://github.com/tenseijs/tensei/commit/b5595bef28486c2c8191f3b8c69f519d683ea22f))
+
+
+
+
+
 ## [0.11.3](https://github.com/tenseijs/tensei/compare/v0.11.2...v0.11.3) (2022-01-05)
 
 **Note:** Version bump only for package @tensei/main